Create SidebarItem component
